% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-object.R
\name{is_rs}
\alias{is_rs}
\title{Test whether an object is an ocular object}
\usage{
is_rs(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical.
}
\description{
Test whether an object is an ocular object
}
\examples{
is_rs(list())
is_rs(structure(list(), class = "ocular"))
}
